+++ src/hooks/useBalanceStore.ts
@@ -1,0 +1,7 @@
+import { useWalletStore } from './useWalletStore';
+
+export const useBalanceStore = () => {
+  const walletStore = useWalletStore();
+
+  return walletStore.balanceStore;
+};